requirements.txt
asgiref==3.6.0
Django==4.1.7
psycopg2-binary==2.9.5
djangorestframework~=3.14.0
